Moved game-container to center, updated folder structure
