+++ index.html
@@ -3,7 +3,12 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="icon" href="./images/favicon.ico" type="image/x-icon" />
+    <link
+      rel="icon"
+      href="./resources/images/favicon.ico"
+      type="image/x-icon"
+    />
+    <link rel="stylesheet" href="./resources/styles.css" />
     <title>Sepeli</title>
   </head>
   <body>
