+++ resources/images/favicon.ico
