+++ resources/images/player_walkA6.png
